Remove firestore.ts entirely
