+++ src/firebase-app.ts
@@ -28,7 +28,7 @@ import { Storage } from './storage/storage';
 import { Database } from '@firebase/database';
 import { DatabaseService } from './database/database';
 import { Firestore } from '@google-cloud/firestore';
-import { FirestoreService } from './firestore/firestore';
+import { FirestoreService } from './firestore/firestore-internal';
 import { InstanceId } from './instance-id/instance-id';
 
 import { ProjectManagement } from './project-management/project-management';
@@ -339,7 +339,7 @@ export class FirebaseApp {
 
   public firestore(): Firestore {
     const service: FirestoreService = this.ensureService_('firestore', () => {
-      const firestoreService: typeof FirestoreService = require('./firestore/firestore').FirestoreService;
+      const firestoreService: typeof FirestoreService = require('./firestore/firestore-internal').FirestoreService;
       return new firestoreService(this);
     });
     return service.client;
